Change README.md, add "name" for Parameters, move Page Objects to "java" directory
