+++ README.md
@@ -1,1 +1,19 @@
+# Sprint_4
 
+Репозиторий хранит исходный код для автоматизированных тестов на Java для учебного проекта Яндекс.Практикума из Спринта №4.
+
+[Учебный сайт Яндекс.Самокат](https://qa-scooter.praktikum-services.ru)
+
+## Технологии
+
+В проекте используется следующий стек:
+1. Java 11
+2. JUnit 4.13.2
+3. Selenium-Java 3.141.59
+4. Boni García WebDriverManager 4.4.3
+5. Maven 3.9.0
+
+## Данные об учащемся
+- Сергеева Александра Витальевна
+- Курс "Автоматизатор тестирования на Java"
+- Поток 19
