+++ src/main/java/pageObjects/OrderPage.java
@@ -1,4 +1,4 @@
-package ui.pages;
+package pageObjects;
 
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
