+++ src/test/java/ui/MainPageLinksTests.java
@@ -1,7 +1,7 @@
 package ui;
 
 import io.github.bonigarcia.wdm.WebDriverManager;
-import ui.pages.MainPage;
+import pageObjects.MainPage;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
